refactor: 메인페이지 및 공통 컴포넌트 폴더 구조 수정
