+++ src/app/_components/Header.tsx
@@ -59,10 +59,10 @@ const Header = (): React.JSX.Element => {
         </h1>
         <ul className={styles.navi}>
           <li>
-            <Link href="/category">크리스마스</Link>
+            <Link href="/">크리스마스</Link>
           </li>
           <li>
-            <Link href="/meetup">촛불 모임</Link>
+            <Link href="/">촛불모임</Link>
           </li>
           <li>
             <Link href="/developers">개발자</Link>
