+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import type { Metadata } from 'next'
 import './globals.scss'
-import Header from '../components/layout/Header'
-import Footer from '../components/layout/Footer'
+import Header from './_components/Header'
+import Footer from './_components/Footer'
 import ScrollBtn from './(common)/_components/ScrollBtn'
 
 export const metadata: Metadata = {
